requirements.txt
Flask
requests
google-api-python-client
youtube-transcript-api
mysql-connector-python
wikitextparser
pattern
langdetect